ARUHA-267: extracted rebalance tests to separate class
