+++ src/acceptance-test/java/de/zalando/aruha/nakadi/webservice/hila/HilaAT.java
@@ -12,20 +12,13 @@ import org.junit.Before;
 import org.junit.Test;
 
 import java.io.IOException;
-import java.util.List;
-import java.util.Set;
 
-import static com.google.common.collect.Sets.intersection;
 import static de.zalando.aruha.nakadi.utils.TestUtils.waitFor;
 import static de.zalando.aruha.nakadi.webservice.hila.StreamBatch.singleEventBatch;
 import static de.zalando.aruha.nakadi.webservice.utils.NakadiTestUtils.commitCursors;
-import static de.zalando.aruha.nakadi.webservice.utils.NakadiTestUtils.createBusinessEventTypeWithPartitions;
 import static de.zalando.aruha.nakadi.webservice.utils.NakadiTestUtils.createEventType;
 import static de.zalando.aruha.nakadi.webservice.utils.NakadiTestUtils.createSubscription;
-import static de.zalando.aruha.nakadi.webservice.utils.NakadiTestUtils.publishBusinessEventWithUserDefinedPartition;
 import static de.zalando.aruha.nakadi.webservice.utils.NakadiTestUtils.publishEvent;
-import static java.util.stream.Collectors.toList;
-import static java.util.stream.Collectors.toSet;
 import static java.util.stream.IntStream.range;
 import static java.util.stream.IntStream.rangeClosed;
 import static org.apache.http.HttpStatus.SC_OK;
@@ -92,66 +85,6 @@ public class HilaAT extends BaseAT {
         assertThat(commitResult, equalTo(SC_OK));
     }
 
-    @Test(timeout = 30000)
-    public void whenRebalanceThenPartitionsAreEquallyDistributedAndCommittedOffsetsAreConsidered() throws Exception {
-        eventType = createBusinessEventTypeWithPartitions(8);
-        subscription = createSubscription(ImmutableSet.of(eventType.getName()));
-
-        // write 5 events to each partition
-        range(0, 40)
-                .forEach(x -> publishBusinessEventWithUserDefinedPartition(
-                        eventType.getName(), "blah" + x, String.valueOf(x % 8)));
-
-        // create a session
-        final TestStreamingClient clientA = TestStreamingClient
-                .create(URL, subscription.getId(), "")
-                .start();
-        waitFor(() -> assertThat(clientA.getBatches(), hasSize(40)));
-
-        // check that we received 5 events for each partitions
-        range(0, 8).forEach(partition ->
-                assertThat(
-                        clientA.getBatches().stream()
-                                .filter(batch -> batch.getCursor().getPartition().equals(String.valueOf(partition)))
-                                .count(),
-                        equalTo(5L)));
-
-        // commit what we consumed
-        final List<Cursor> cursors = range(0, 8)
-                .boxed()
-                .map(partition -> new Cursor(String.valueOf(partition), "4"))
-                .collect(toList());
-        commitCursors(subscription.getId(), cursors);
-
-        // create second session for the same subscription
-        final TestStreamingClient clientB = TestStreamingClient
-                .create(URL, subscription.getId(), "")
-                .start();
-
-        // wait for rebalance process to start
-        Thread.sleep(1000);
-
-        // write 5 more events to each partition
-        range(0, 40)
-                .forEach(x -> publishBusinessEventWithUserDefinedPartition(
-                        eventType.getName(), "blahAgain" + x, String.valueOf(x % 8)));
-
-        // wait till all event arrive
-        waitFor(() -> assertThat(clientB.getBatches(), hasSize(20)));
-        waitFor(() -> assertThat(clientA.getBatches(), hasSize(60)));
-
-        // check that only half of partitions was streamed to client A after rebalance
-        final Set<String> clientAPartitionsAfterRebalance = getUniquePartitionsStreamedToClient(clientA, 40, 60);
-        assertThat(clientAPartitionsAfterRebalance, hasSize(4));
-
-        // check that only half of partitions was streamed to client B
-        final Set<String> clientBPartitions = getUniquePartitionsStreamedToClient(clientB);
-        assertThat(clientBPartitions, hasSize(4));
-
-        // check that different partitions were streamed to different clients
-        assertThat(intersection(clientAPartitionsAfterRebalance, clientBPartitions), hasSize(0));
-    }
-
     @Test(timeout = 15000)
     public void whenWindowSizeIsSetItIsConsidered() throws Exception {
 
@@ -163,13 +96,13 @@ public class HilaAT extends BaseAT {
 
         waitFor(() -> assertThat(client.getBatches(), hasSize(5)));
 
-        final Cursor lastBatchCursor = client.getBatches().get(client.getBatches().size() - 1).getCursor();
-        commitCursors(subscription.getId(), ImmutableList.of(lastBatchCursor));
+        Cursor cursorToCommit = client.getBatches().get(4).getCursor();
+        commitCursors(subscription.getId(), ImmutableList.of(cursorToCommit));
 
         waitFor(() -> assertThat(client.getBatches(), hasSize(10)));
 
-        final Cursor cursor = client.getBatches().get(client.getBatches().size() - 4).getCursor();
-        commitCursors(subscription.getId(), ImmutableList.of(cursor));
+        cursorToCommit = client.getBatches().get(6).getCursor();
+        commitCursors(subscription.getId(), ImmutableList.of(cursorToCommit));
 
         waitFor(() -> assertThat(client.getBatches(), hasSize(12)));
     }
@@ -199,7 +132,7 @@ public class HilaAT extends BaseAT {
         waitFor(() -> assertThat(client.getBatches(), hasSize(1)));
 
         // to check that stream_timeout works we need to commit everything we consumed, in other case
-        // Nakadi will wait for commit the amount of time defined by commit_timeout
+        // Nakadi will first wait till commit_timeout exceeds
         final Cursor lastBatchCursor = client.getBatches().get(client.getBatches().size() - 1).getCursor();
         commitCursors(subscription.getId(), ImmutableList.of(lastBatchCursor));
 
@@ -225,20 +158,6 @@ public class HilaAT extends BaseAT {
 
         assertThat(client.getBatches().get(2).getEvents(), hasSize(2));
         assertThat(client.getBatches().get(2).getCursor().getOffset(), is("11"));
-    }
-
-    private Set<String> getUniquePartitionsStreamedToClient(final TestStreamingClient client) {
-        return getUniquePartitionsStreamedToClient(client, 0, client.getBatches().size());
-    }
-
-    private Set<String> getUniquePartitionsStreamedToClient(final TestStreamingClient client, final int fromBatch,
-                                                            final int toBatch) {
-        return client.getBatches()
-                .subList(fromBatch, toBatch)
-                .stream()
-                .map(batch -> batch.getCursor().getPartition())
-                .distinct()
-                .collect(toSet());
     }
 
 }
